Refactors readers
